Rename routes to routers
